+++ constcheck/_core.py
@@ -55,6 +55,9 @@ def _get_strings(textio: _t.TextIO) -> _TokenList:
             # prevent appending to previous bracketed multiline str
             split = True
 
+        if any([ttext.islsqb(), ttext.isrsqb()]):
+            split = True
+
         if ttype.isstr() and not ttext.isdoc(prev_ttext, prev_ttype):
             ttext = ttext.strip().lstrip()